#62 Display current preview time position
